refactor(fused-moe): split runtime contracts into dataclass modules
Move fused MoE runtime contracts into dedicated dataclass modules, keep the token_dispatcher facade backward compatible, and add a leaf router_input contract to break import cycles.

Also restore legacy builder compatibility for activation and fusion flags, then update focused unit tests to cover the refactor under the current local torch_npu environment.
