+++ vllm_ascend/_310p/quantization/methods/w8a8_dynamic.py
@@ -23,7 +23,7 @@ from vllm.config import get_current_vllm_config
 from vllm.distributed import get_ep_group
 
 from vllm_ascend.ascend_forward_context import _EXTRA_CTX
-from vllm_ascend.ops.fused_moe.moe_runtime_args import build_fused_experts_input
+from vllm_ascend.ops.fused_moe.dataclass.fused_experts import build_fused_experts_input
 from vllm_ascend.ops.fused_moe.routed_experts import AscendRoutedExperts
 from vllm_ascend.quantization.methods.base import AscendMoEScheme, QuantType
 from vllm_ascend.utils import maybe_trans_nz
@@ -105,7 +105,6 @@ class AscendW8A8DynamicFusedMoEMethod310(AscendMoEScheme):
                 pertoken_scale=layer.ascend_pertoken_scale,
                 w1_scale=layer.w13_weight_scale,
                 w2_scale=layer.w2_weight_scale,
-
             ),
         )
         return final_hidden_states
