+++ vllm_ascend/ops/fused_moe/dataclass/prepare_finalize.py
@@ -1,0 +1,30 @@
+#
+# Copyright (c) 2025 Huawei Technologies Co., Ltd. All Rights Reserved.
+# Copyright 2023 The vLLM team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+from dataclasses import dataclass
+
+import torch
+
+
+@dataclass(frozen=True, slots=True)
+class MoEPrepareOutput:
+    """Typed output from prepare stage."""
+
+    hidden_states: torch.Tensor
+    router_logits: torch.Tensor
+    mc2_mask: torch.Tensor | None
+    padded_hidden_states_shape: torch.Size | None
+    pertoken_scale: torch.Tensor | None = None
