+++ vllm_ascend/ops/fused_moe/moe_comm_method.py
@@ -26,15 +26,13 @@ from vllm_ascend.ascend_config import get_ascend_config
 from vllm_ascend.ascend_forward_context import _EXTRA_CTX, _MEGA_MOE_SUPPORTED, MoECommType
 from vllm_ascend.distributed.parallel_state import get_mc2_group
 from vllm_ascend.ops.fused_moe import moe_utils
-from vllm_ascend.ops.fused_moe.moe_mlp import unified_apply_mlp
-from vllm_ascend.ops.fused_moe.moe_runtime_args import (
+from vllm_ascend.ops.fused_moe.dataclass.moe_mlp import MoEMlpComputeInput, build_mlp_compute_input
+from vllm_ascend.ops.fused_moe.dataclass.prepare_finalize import MoEPrepareOutput
+from vllm_ascend.ops.fused_moe.dataclass.token_dispatcher import (
     MoEFusedExpertsInput,
-    MoEMlpComputeInput,
-    MoEPrepareOutput,
-    build_mlp_compute_input,
     build_token_dispatch_input,
 )
-from vllm_ascend.ops.fused_moe.moe_utils import enable_fusion_gmmswigluquant
+from vllm_ascend.ops.fused_moe.moe_mlp import unified_apply_mlp
 from vllm_ascend.ops.fused_moe.prepare_finalize import (
     PrepareAndFinalize,
     PrepareAndFinalizeWithAll2All,
@@ -147,6 +145,7 @@ class MoECommMethod(ABC):
         mlp_compute_input = build_mlp_compute_input(
             fused_experts_input=fused_experts_input,
             token_dispatch_output=token_dispatch_output,
+            moe_config=self.moe_config,
         )
 
         mlp_output, before_gmm2_evt = self._apply_mlp(mlp_compute_input)
