+++ vllm_ascend/ops/fused_moe/routed_experts.py
@@ -34,9 +34,9 @@ from vllm_ascend.ascend_forward_context import _EXTRA_CTX, MoECommType
 from vllm_ascend.eplb.adaptor.vllm_adaptor import VllmEplbAdaptor
 from vllm_ascend.eplb.core.eplb_utils import init_eplb_config
 from vllm_ascend.lora.fused_moe import sync_lora_context
+from vllm_ascend.ops.fused_moe.dataclass.fused_experts import build_fused_experts_input
 from vllm_ascend.ops.fused_moe.eplb import record_local_expert_load
 from vllm_ascend.ops.fused_moe.moe_comm_method import AllGatherCommImpl, FusedExpertsResult
-from vllm_ascend.ops.fused_moe.moe_runtime_args import build_fused_experts_input
 from vllm_ascend.ops.fused_moe.moe_utils import get_moe_num_logical_experts
 from vllm_ascend.ops.fused_moe.shared_experts import FusedMoEEvents
 from vllm_ascend.quantization.quant_type import QuantType
@@ -111,10 +111,6 @@ class AscendUnquantizedFusedMoEMethod(UnquantizedFusedMoEMethod):
         shared_experts: SharedExperts | None,
         shared_experts_input: torch.Tensor | None,
     ) -> torch.Tensor:
-        activation = getattr(layer, "activation", "silu")
-        if getattr(layer, "swigluoai_uninterleave", False):
-            activation = "swigluoai_uninterleave"
-
         moe_comm_method = _EXTRA_CTX.moe_comm_method
         w13_weight_list = getattr(layer, "w13_weight_list", None)
         w2_weight_list = getattr(layer, "w2_weight_list", None)
@@ -457,7 +453,7 @@ class AscendRoutedExperts(RoutedExperts):  # type: ignore[no-redef]
                 capturer = getattr(self, "_ascend_routed_experts_capturer", None)
                 if capturer is not None:
                     capturer.capture(layer_id=self.layer_id, topk_ids=topk_ids)
-        except Exception as e:
+        except Exception:
             pass
 
         num_shared_experts = self.n_shared_experts
